services/users/requirements.txt
fastapi
prometheus_fastapi_instrumentator
uvicorn
psycopg2
pydantic
pydantic[email]
typing
python-dotenv
python-multipart
python-keycloak
